Optimize returning a value in a0
lcc doesn't seem to construct the return value in the return register, instead it just copies it at the end of the function. Therefore, it's safe to assume that there is no useful information stored in a0 at the time of returning.
This means that we can just copy the return value to a0 without performing extra checks and copies.