Add new functions to Interface Record for smarter register allocation
void info_var(float ref, int addressed);
int var_register(float ref);
For each local variable, info_var is called so that the backend is able to gather information about the variables of the function.
Then, for each non-addressed variable, var_register is called. The backend returns 1 if the variable should be stored in a register.